Move logic out of watch package
